Configuration Screen: use int consistently for variable expressions (#66)
Boundaries and increments were expressed as double, even tough all
constants are integer values. Also avoid mixing int/char.

+++ wavemon.h
@@ -174,7 +174,7 @@ struct conf_item {
 	int	*dep;		/* dependency */
 	bool	hidden;		/* only allow to set via wavemonrc */
 
-	double	min,		/* value boundaries */
+	int	min,		/* value boundaries */
 		max,
 		inc;		/* increment for value changes */
